Revert "csitool: enable 5 GHz operation for experiments"
This reverts commits a74e0563aa2a7f61b668daf80f1148f0fc967b04 and
294a30c83358a955c40d7d7f0fc5927fe90de1c1. The regulatory settings
read from the EEPROM are used not just by the driver, but also by
the firmware. While the driver may be able to work differently by
ignoring the EEPROM settings, this approach is not ideal, because
it makes it less transparent to the user what is happening when an
operation such as IBSS beaconing is blocked outside of the driver.

Signed-off-by: David Ward <david.ward@ll.mit.edu>